* commit 'b67486580d0d641a286ae59bd5e75bc299459ec3':
  修复端口状态扫描慢的问题
  修复端口状态扫描慢的问题

+++ dmestore-service/src/main/java/com/dmeplugin/dmestore/model/VmfsDataInfo.java
@@ -82,7 +82,7 @@ public class VmfsDataInfo {
     /**
      * dme中时延 单位ms
      **/
-    Integer latency;
+    Float latency;
     /**
      * dme中卷ID
      **/
@@ -240,11 +240,11 @@ public class VmfsDataInfo {
         this.writeResponseTime = writeResponseTime;
     }
 
-    public Integer getLatency() {
+    public Float getLatency() {
         return latency;
     }
 
-    public void setLatency(Integer latency) {
+    public void setLatency(Float latency) {
         this.latency = latency;
     }
